Added mapDispatchToProps as second optional parameter into connect function when exporting CoursesPage.js.
Defined mapDispatchToProps to get a dispatch and return a object which contains the result of calling the dispatch with the action to create the course.

Modified onclicksave deleting dispatch(courseActions because of we are using the dispatcher into the map itself.

Removed the proptype to rquire for dispatcher because connect doesn't need it.